+++ Reddit/Easy/integer_complexity_1/src/lib.rs
@@ -1,5 +1,7 @@
 
 pub fn smallest_factor_sum(n: i32) -> i32 {
+    // TODO: prime test here
+
     println!("got: {}", n);
 
     0