+++ Reddit/Easy/integer_complexity_1/src/tests.rs
@@ -1,5 +1,10 @@
+use super::*;
 
 #[test]
-fn stub_test() {
-    assert_eq!(1, 1);
+fn basic_factor_test() {
+    assert_eq!(838, smallest_factor_sum(12345));
+    assert_eq!(7, smallest_factor_sum(12));
+    assert_eq!(43, smallest_factor_sum(456));
+    assert_eq!(4568, smallest_factor_sum(4567));
 }
+